+++ builder/googlecompute/step_create_ssh_key_test.go
@@ -2,6 +2,9 @@ package googlecompute
 
 import (
 	"github.com/mitchellh/multistep"
+
+	"io/ioutil"
+	"os"
 	"testing"
 )
 
@@ -25,5 +28,36 @@ func TestStepCreateSSHKey(t *testing.T) {
 	}
 	if _, ok := state.GetOk("ssh_public_key"); !ok {
 		t.Fatal("should have key")
+	}
+}
+
+func TestStepCreateSSHKey_debug(t *testing.T) {
+	tf, err := ioutil.TempFile("", "packer")
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+	tf.Close()
+
+	state := testState(t)
+	step := new(StepCreateSSHKey)
+	step.Debug = true
+	step.DebugKeyPath = tf.Name()
+
+	defer step.Cleanup(state)
+
+	// run the step
+	if action := step.Run(state); action != multistep.ActionContinue {
+		t.Fatalf("bad action: %#v", action)
+	}
+
+	// Verify that we have a public/private key
+	if _, ok := state.GetOk("ssh_private_key"); !ok {
+		t.Fatal("should have key")
+	}
+	if _, ok := state.GetOk("ssh_public_key"); !ok {
+		t.Fatal("should have key")
+	}
+	if _, err := os.Stat(tf.Name()); err != nil {
+		t.Fatalf("err: %s", err)
 	}
 }